few enhancements
