+++ src/Processing/Producer.cs
@@ -6,140 +6,35 @@ using System.Text;
 
 namespace PivotalServices.RabbitMQ.Messaging;
 
-public interface IProducer<T> : IDisposable
+public interface IProducer
 {
-    void Send(OutboundMessage<T> message);
     void Close();
 }
 
-public class Producer<T> : IProducer<T>
+public interface IProducer<T> : IProducer, IDisposable
 {
-    private readonly IOptions<ServiceConfiguration> serviceConfigurationOptions;
-    protected QueueConfiguration queueConfiguration;
-    protected DlxQueueConfiguration deadLetterQueueConfiguration;
+    void Send(OutboundMessage<T> message);
+}
+
+public class Producer<T> : Initializer<T>, IProducer<T>
+{
     ILogger<Producer<T>> logger;
-    protected IConnection connection;
-    protected IModel channel;
-    protected IBasicProperties basicProperties;
-    protected IConnectionFactory factory;
-    protected string exchangeName;
-    protected string queueName;
-    protected List<string> routingKeysFromConfiguration;
     protected bool disposedValue = false;
     private bool isQueueBindingRequired;
 
     public Producer(IOptions<ServiceConfiguration> serviceConfigurationOptions,
-                IOptionsMonitor<QueueConfiguration> queueConfigurationOptions,
-                IOptionsMonitor<DlxQueueConfiguration> deadLetterQueueConfigurationOptions,
-                ILogger<Producer<T>> logger)
+                    IOptionsMonitor<QueueConfiguration> queueConfigurationOptions,
+                    IOptionsMonitor<DlxQueueConfiguration> deadLetterQueueConfigurationOptions,
+                    ILogger<Producer<T>> logger)
+            : base(serviceConfigurationOptions, 
+                    queueConfigurationOptions,
+                    deadLetterQueueConfigurationOptions,
+                    logger)
+        
     {
-        var optionsName = typeof(T).Name;
-        this.serviceConfigurationOptions = serviceConfigurationOptions;
-        this.queueConfiguration = queueConfigurationOptions.Get(optionsName);
-        this.deadLetterQueueConfiguration = deadLetterQueueConfigurationOptions.Get(optionsName);
         this.logger = logger;
-        routingKeysFromConfiguration = queueConfiguration.RoutingKeysCsv == null 
-                                            ? new List<string>() 
-                                            : queueConfiguration.RoutingKeysCsv.Split(',').ToList();
         isQueueBindingRequired = queueConfiguration.ExchangeType != ExchangeType.Topic
                                     && queueConfiguration.ExchangeType != ExchangeType.Fanout;
-        Initialize();
-    }
-
-    private void Initialize()
-    {
-        logger.LogInformation($"Initializing Producer<{typeof(T).Name}> - {queueConfiguration.QueueName}");
-        logger.LogDebug($"Initializing Producer<{typeof(T).Name}> with Configuration- {JsonConvert.SerializeObject(queueConfiguration)}");
-
-        factory = CreateConnectionFactory();
-        connection = factory.CreateConnection();
-        channel = connection.CreateModel();
-
-        exchangeName = queueConfiguration.ExchangeName;
-        var properties = new Dictionary<string, object>();
-
-        channel.ExchangeDeclare(exchange: exchangeName,
-                                type: queueConfiguration.ExchangeType);
-
-        if (!string.IsNullOrEmpty(deadLetterQueueConfiguration.ExchangeNamePrefix)
-        && !string.IsNullOrEmpty(deadLetterQueueConfiguration.QueueNamePrefix))
-        {
-            CreateDeadLetterExchange(properties);
-        }
-
-        properties["x-max-length"] = queueConfiguration.MaximumQueueLength;
-        properties["x-max-priority"] = queueConfiguration.MaxPriority;
-
-        if (isQueueBindingRequired)
-        {
-            DeclareQueue(properties);
-        }
-
-        channel.BasicQos(prefetchSize: 0,
-                         prefetchCount: queueConfiguration.PrefetchCount,
-                         global: false);
-
-        basicProperties = channel.CreateBasicProperties();
-        basicProperties.Expiration = (queueConfiguration.MessageExpirationInSeconds * 1000).ToString();
-        basicProperties.Persistent = queueConfiguration.IsPersistent;
-    }
-
-    protected void DeclareQueue(Dictionary<string, object> properties)
-    {
-        queueName = channel.QueueDeclare(queue: queueConfiguration.QueueName,
-                                         durable: queueConfiguration.IsDurable,
-                                         exclusive: queueConfiguration.IsExclusive,
-                                         autoDelete: queueConfiguration.AutoDelete,
-                                         arguments: properties
-                                         ).QueueName;
-
-        if (!routingKeysFromConfiguration.Any())
-        {
-            channel.QueueBind(queue: queueName,
-                              exchange: exchangeName,
-                              routingKey: string.Empty,
-                              arguments: null);
-        }
-        else
-        {
-            foreach (var routingKey in routingKeysFromConfiguration)
-            {
-                channel.QueueBind(queue: queueName,
-                              exchange: exchangeName,
-                              routingKey: routingKey.Trim(),
-                              arguments: null);
-            }
-        }
-    }
-
-    private void CreateDeadLetterExchange(Dictionary<string, object> properties)
-    {
-        var dlxExchangeName = $"{deadLetterQueueConfiguration.ExchangeNamePrefix}-DLX";
-        var dlxQueueName = $"{deadLetterQueueConfiguration.QueueNamePrefix}-DLQ";
-        var dlxProperties = new Dictionary<string, object>();
-        channel.QueueDeclare(queue: dlxQueueName,
-                             durable: false,
-                             exclusive: false,
-                             autoDelete: false,
-                             arguments: dlxProperties);
-        channel.QueueBind(queue: dlxQueueName,
-                          exchange: dlxExchangeName,
-                          routingKey: string.Empty,
-                          arguments: null);
-        properties["x-dead-letter-exchange"] = dlxExchangeName;
-
-    }
-
-    protected virtual IConnectionFactory CreateConnectionFactory()
-    {
-        return new ConnectionFactory
-        {
-            HostName = serviceConfigurationOptions.Value.HostName,
-            VirtualHost = serviceConfigurationOptions.Value.Vhost,
-            UserName = serviceConfigurationOptions.Value.Username,
-            Password = serviceConfigurationOptions.Value.Password,
-            AutomaticRecoveryEnabled = true,
-        };
     }
 
     public void Send(OutboundMessage<T> message)
@@ -167,10 +62,10 @@ public class Producer<T> : IProducer<T>
         {
             if (isQueueBindingRequired)
             {
-                channel.QueueBind(queueName, exchangeName, routingKey.Trim(), null);
+                channel.QueueBind(queueName, queueConfiguration.ExchangeName, routingKey.Trim(), null);
             }
 
-            channel.BasicPublish(exchangeName, routingKey.Trim(), basicProperties, messageBody);
+            channel.BasicPublish(queueConfiguration.ExchangeName, routingKey.Trim(), basicProperties, messageBody);
             logger.LogDebug($"Sent message {JsonConvert.SerializeObject(message)}");
             logger.LogInformation($"Sent a message with CorreleationId {message.CorrelationId}");
         }
@@ -178,6 +73,9 @@ public class Producer<T> : IProducer<T>
 
     public void Close()
     {
+        if(!connection.IsOpen)
+            return;
+
         channel.Close();
 
         if (connection != null && connection.IsOpen)
